move config reading to separate package internal/config
